pyproject.toml
[project]
name = "make-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=0.7.501",
    "google-generativeai>=0.8.5",
    "googletrans==4.0.0rc1",
    "langdetect>=1.0.9",
    "python-dotenv>=1.1.0",
]
